Add support for sort keys on dynamo table
Use Case: Users of this library may have existing dynamo tables that have a composite (hash and range) key and wish to use that existing table for the locking as well. Alternatively, users may wish to use a single dynamo table, but have an additional sort key to share the table between services, each with a unique value for the sort key.

Adds support for a sort key (range) on all dynamodb interactions including GetItem, UpdateItem, and CreateTable. Adds tests for client using the new sort keys. The tests are just a copy of the original tests for client, adding the additional sort key arguments.

Signed-off-by: Joe Adams <github@joeadams.io>