Use `internalClient` internally instead of `Client`
Both `Client` and the eventual `ClientWithSortKey` are wrappers around
`internalClient`, which is what the current `Client` is. This allows
us to keep much of the shared functionality the same between the two
types, and only changing a handful of method signatures where a sort key
must be present.

This commit shouldn't functionally change anything, as it's simply
updating method signatures.