Make the key variable naming scheme consistent.
Currently, the variable naming scheme for keys is to call the partition
key's name itself `partitionKeyName`, and its value `key`. Since we'll
be introducing a sort key, I propose we rename all relevant `key`
variables to `partitionKey` instead. This way, later on we can add in
a `sortKeyName` and `sortKey` pair.

There's a few places where `key` is used as a variable name which I left
untouched (eg within `releaseLock`). These keys will eventually be a
combination of both the partition and sort keys, so it makes sense to
keep their name generic.

Additionally, the `getLockOptions` struct had a field named
`partitionKeyName`, but which was being used to store a simple partition
key. As such, I renamed it to `partitionKey`, to be consistent with the
other uses of the name `partitionKey`.