dynamolock/v2: improve RVN entropy
This commit reduces the likelihood of RVN collision by adding a
dimension of time to the RVN random value. It uses nanosecond resolution
(or as close to nanosecond as possible depending on the platform) so
that even if the string part of the RVN collides, the time part of
the RVN will greatly reduce the likelihood of two same RVNs happening
in the same nanosecond (or close to the order of magnitude of nanosecond).